summarizing the data for Gradient Boosting
Percussion Pressure: Ranges from 0.213 to 201.141, with a mean of 182.33 and a standard deviation of 19.26.
Feed Pressure: Ranges from 0.342 to 93.844, with a mean of 81.15 and a standard deviation of 17.28.
Flush Air Pressure: Ranges from 2.173 to 11.460, with a mean of 7.97 and a standard deviation of 1.19.
Rotation Pressure: Ranges from 29.829 to 165.846, with a mean of 53.41 and a standard deviation of 5.82.
Dampening Pressure: Ranges from 35.467 to 81.053, with a mean of 65.87 and a standard deviation of 9.16.
Penetration Rate (Target): Ranges from 0.350 to 19.046, with a mean of 2.12 and a standard deviation of 0.69.